Refactor: move MusicOptionsDialog to subfolder
